requirements.txt
pipdeptree==0.10.1
pkg-resources==0.0.0
pylint==1.7.2
python-decouple==3.1
requests==2.18.4
setuptools==36.5.0
wheel==0.30.0
